requirements.txt
langchain
langchain-groq
python-dotenv
langchain-community
langgraph
arxiv
wikipedia
Flask
SpeechRecognition  # If using voice processing server-side
pyttsx3 